bancolombia-rag/scraper_service/requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1
